Setting up and tearing down a database

+++ Pluralsight.OutsideInTDD.WebAPI/Domain/JournalEntry.cs
@@ -5,7 +5,7 @@ namespace Pluralsight.OutsideInTDD.WebAPI.Domain
     public class JournalEntry
     {
         public DateTime Time { get; set; }
-        public decimal Distance { get; set; }
+        public int Distance { get; set; }
         public TimeSpan Duration { get; set; }
     }
 }
